+++ module-core/src/main/java/com/geekhome/common/IHardwareManagerAdapter.java
@@ -9,7 +9,7 @@ import java.util.Calendar;
 
 public interface IHardwareManagerAdapter extends IMonitorable, INamedObject {
     void discover(InputPortsCollection<Boolean> digitalInputPorts, OutputPortsCollection<Boolean> digitalOutputPorts,
-                  InputPortsCollection<Integer> powerInputPorts, OutputPortsCollection<Integer> powerOutputPorts,
+                  InputPortsCollection<Double> powerInputPorts, OutputPortsCollection<Integer> powerOutputPorts,
                   InputPortsCollection<Double> temperaturePorts, TogglePortsCollection togglePorts,
                   InputPortsCollection<Double> humidityPorts, InputPortsCollection<Double> luminosityPorts) throws DiscoveryException;
     void refresh(Calendar now) throws Exception;
