+++ module-gree/src/main/java/com/geekhome/shellymodule/GreeAdapter.java
@@ -50,7 +50,7 @@ class GreeAdapter extends NamedObject implements IHardwareManagerAdapter {
     @Override
     public void discover(final InputPortsCollection<Boolean> digitalInputPorts,
                          final OutputPortsCollection<Boolean> digitalOutputPorts,
-                         final InputPortsCollection<Integer> powerInputPorts,
+                         final InputPortsCollection<Double> powerInputPorts,
                          final OutputPortsCollection<Integer> powerOutputPorts,
                          final InputPortsCollection<Double> temperaturePorts,
                          final TogglePortsCollection togglePorts,
