+++ module-mqtt/src/main/java/com/geekhome/mqttmodule/MqttAdapter.java
@@ -51,7 +51,7 @@ class MqttAdapter extends NamedObject implements IHardwareManagerAdapter, MqttCa
     @Override
     public void discover(final InputPortsCollection<Boolean> digitalInputPorts,
                            final OutputPortsCollection<Boolean> digitalOutputPorts,
-                           final InputPortsCollection<Integer> powerInputPorts,
+                           final InputPortsCollection<Double> powerInputPorts,
                            final OutputPortsCollection<Integer> powerOutputPorts,
                            final InputPortsCollection<Double> temperaturePorts,
                            final TogglePortsCollection togglePorts,
