+++ module-onewire/src/main/java/com/geekhome/onewiremodule/OneWireAdapter.java
@@ -48,7 +48,7 @@ class OneWireAdapter extends SerialAdapterBase {
 
     @Override
     public void discover(InputPortsCollection<Boolean> digitalInputPorts, OutputPortsCollection<Boolean> digitalOutputPorts,
-                         InputPortsCollection<Integer> powerInputPorts, OutputPortsCollection<Integer> powerOutputPorts,
+                         InputPortsCollection<Double> powerInputPorts, OutputPortsCollection<Integer> powerOutputPorts,
                          InputPortsCollection<Double> temperaturePorts, TogglePortsCollection togglePorts,
                          InputPortsCollection<Double> humidityPorts, InputPortsCollection<Double> luminosityPorts) throws DiscoveryException {
         _switchesAsSensedInputs.clear();
